refactor(kotlin): refactor quick_sort.kt files.
